requirements.txt
atproto
python-dotenv
colorama